+++ src/transactions/instances/index.ts
@@ -10,3 +10,4 @@ export * from "./signedTransaction";
 export * from "./transactionArgument";
 export * from "./transactionPayload";
 export * from "./simpleTransaction";
+export * from "./multiAgentTransaction";