+++ src/utils/apiEndpoints.ts
@@ -5,7 +5,6 @@ export const NetworkToIndexerAPI: Record<string, string> = {
   mainnet: "https://api.mainnet.aptoslabs.com/v1/graphql",
   testnet: "https://api.testnet.aptoslabs.com/v1/graphql",
   devnet: "https://api.devnet.aptoslabs.com/v1/graphql",
-  randomnet: "https://indexer-randomnet.hasura.app/v1/graphql",
   local: "http://127.0.0.1:8090/v1/graphql",
 };
 
@@ -13,7 +12,6 @@ export const NetworkToNodeAPI: Record<string, string> = {
   mainnet: "https://api.mainnet.aptoslabs.com/v1",
   testnet: "https://api.testnet.aptoslabs.com/v1",
   devnet: "https://api.devnet.aptoslabs.com/v1",
-  randomnet: "https://fullnode.random.aptoslabs.com/v1",
   local: "http://127.0.0.1:8080/v1",
 };
 
@@ -21,7 +19,6 @@ export const NetworkToFaucetAPI: Record<string, string> = {
   mainnet: "https://faucet.mainnet.aptoslabs.com",
   testnet: "https://faucet.testnet.aptoslabs.com",
   devnet: "https://faucet.devnet.aptoslabs.com",
-  randomnet: "https://faucet.random.aptoslabs.com",
   local: "http://127.0.0.1:8081",
 };
 
@@ -29,7 +26,6 @@ export enum Network {
   MAINNET = "mainnet",
   TESTNET = "testnet",
   DEVNET = "devnet",
-  RANDOMNET = "randomnet",
   LOCAL = "local",
   CUSTOM = "custom",
 }
@@ -37,7 +33,6 @@ export enum Network {
 export const NetworkToChainId: Record<string, number> = {
   mainnet: 1,
   testnet: 2,
-  randomnet: 70,
   local: 4,
 };
 
@@ -45,7 +40,6 @@ export const NetworkToNetworkName: Record<string, Network> = {
   mainnet: Network.MAINNET,
   testnet: Network.TESTNET,
   devnet: Network.DEVNET,
-  randomnet: Network.RANDOMNET,
   local: Network.LOCAL,
   custom: Network.CUSTOM,
 };